Refine logging utilities and restore server checks
