+++ src/logging.h
@@ -1,0 +1,17 @@
+#ifndef LOGGING_H
+#define LOGGING_H
+
+#include <winsock2.h>
+#include <windows.h>
+#include <stdbool.h>
+#include <stdint.h>
+
+extern CRITICAL_SECTION g_LogCriticalSection;
+extern bool g_CriticalSectionInitialized;
+extern HANDLE g_LogFile;
+extern UINT32 g_LogLineCount;
+
+void logf(const char *fmt, ...);
+void log_winsock_error(const char *prefix, SOCKET s, int error);
+
+#endif // LOGGING_H
