+++ src/main.c
@@ -8,8 +8,8 @@
 #include <stdint.h>
 #include <stdbool.h>
 #include <string.h>
-#include <stdarg.h>
 #include "MinHook.h"
+#include "logging.h"
 
 // Link required libraries
 #pragma comment(lib, "ws2_32.lib")
@@ -19,72 +19,15 @@
 
 // Configuration constants
 static const char *kServerPath = "Server\\server.dll";
-static const DWORD kMaxLogLines = 500;  // Max lines before rollover
 
 // Thread synchronization and module tracking
-static CRITICAL_SECTION g_LogCriticalSection;
-static bool g_CriticalSectionInitialized = false;
 static bool g_HooksInitialized = false;
-static HANDLE g_LogFile = INVALID_HANDLE_VALUE;
-static UINT32 g_LogLineCount = 0;
 
-// Server.dll module range for caller filtering
 static uintptr_t g_ServerBase = 0;
 static size_t g_ServerSize = 0;
 
-/* -------- Utility functions -------- */
-
-static void ResetLogFile(void) {
-    if (g_LogFile != INVALID_HANDLE_VALUE) {
-        SetFilePointer(g_LogFile, 0, NULL, FILE_BEGIN);
-        SetEndOfFile(g_LogFile);
-        g_LogLineCount = 0;
-    }
-}
-
-static void logf(const char *fmt, ...) {
-    if (!g_CriticalSectionInitialized || g_LogFile == INVALID_HANDLE_VALUE) return;
-
-    EnterCriticalSection(&g_LogCriticalSection);
-
-    // Check for log rollover
-    if (++g_LogLineCount > kMaxLogLines) {
-        ResetLogFile();
-    }
-
-    // Format timestamp
-    SYSTEMTIME st;
-    GetLocalTime(&st);
-    char timestamp[64];
-    snprintf(timestamp, sizeof(timestamp), "[%04d-%02d-%02d %02d:%02d:%02d.%03d] ",
-             st.wYear, st.wMonth, st.wDay, st.wHour, st.wMinute, st.wSecond, st.wMilliseconds);
-
-    // Write timestamp
-    DWORD written;
-    WriteFile(g_LogFile, timestamp, (DWORD)strlen(timestamp), &written, NULL);
-
-    // Format and write message
-    char buffer[512];
-    va_list ap;
-    va_start(ap, fmt);
-    int len = vsnprintf(buffer, sizeof(buffer) - 2, fmt, ap);  // Leave space for \r\n
-    va_end(ap);
-
-    if (len > 0) {
-        // Ensure we have newline
-        if (len < sizeof(buffer) - 2 && buffer[len-1] != '\n') {
-            buffer[len++] = '\n';
-            buffer[len] = '\0';
-        }
-        WriteFile(g_LogFile, buffer, len, &written, NULL);
-        FlushFileBuffers(g_LogFile);  // Ensure immediate write
-    }
-
-    LeaveCriticalSection(&g_LogCriticalSection);
-}
-
 static void InitServerModuleRange(void) {
-    if (g_ServerBase != 0) return;  // Already initialized
+    if (g_ServerBase != 0) return;
 
     HMODULE hServer = GetModuleHandleA("server.dll");
     if (!hServer) {
@@ -108,19 +51,18 @@ static BOOL IsCallerFromServer(void) {
         if (g_ServerBase == 0) return FALSE;
     }
 
-    // Capture the caller's address
     CONTEXT ctx = {0};
     ctx.ContextFlags = CONTEXT_CONTROL;
     RtlCaptureContext(&ctx);
 
     uintptr_t callerAddr;
-    #if defined(_M_X64) || defined(__x86_64__)
+#if defined(_M_X64) || defined(__x86_64__)
     callerAddr = ctx.Rip;
-    #elif defined(_M_IX86) || defined(__i386__)
+#elif defined(_M_IX86) || defined(__i386__)
     callerAddr = ctx.Eip;
-    #else
-    #error "Unsupported architecture"
-    #endif
+#else
+#error "Unsupported architecture"
+#endif
 
     BOOL inRange = (callerAddr >= g_ServerBase && callerAddr < g_ServerBase + g_ServerSize);
 
@@ -130,23 +72,6 @@ static BOOL IsCallerFromServer(void) {
     }
 
     return inRange;
-}
-
-static const char* GetErrorDescription(int errorCode) {
-    switch (errorCode) {
-        case WSAECONNRESET: return "WSAECONNRESET";
-        case WSAECONNABORTED: return "WSAECONNABORTED";
-        case WSAENETDOWN: return "WSAENETDOWN";
-        case WSAETIMEDOUT: return "WSAETIMEDOUT";
-        case WSAENOTCONN: return "WSAENOTCONN";
-        case WSAEWOULDBLOCK: return "WSAEWOULDBLOCK";
-        case WSAEINVAL: return "WSAEINVAL";
-        case WSAENOTSOCK: return "WSAENOTSOCK";
-        case WSAEFAULT: return "WSAEFAULT";
-        case WSAEINTR: return "WSAEINTR";
-        case WSAEMSGSIZE: return "WSAEMSGSIZE";
-        default: return "UNKNOWN";
-    }
 }
 
 /* -------- Original function pointers -------- */
@@ -217,38 +142,14 @@ int WINAPI hook_recv(SOCKET s, char *buf, int len, int flags) {
 
     if (result == SOCKET_ERROR) {
         int error = WSAGetLastError();
-        switch (error) {
-            case WSAEWOULDBLOCK:
-                // Convert WSAEWOULDBLOCK to 0 for server.dll calls
-                logf("[WS2 HOOK] recv: WSAEWOULDBLOCK converted to 0 (server.dll caller)");
-                WSASetLastError(NO_ERROR);
-                return 0;
-
-            case WSAECONNRESET:
-                logf("[WS2 HOOK] recv: Connection reset by peer (WSAECONNRESET) on socket %u", (unsigned)s);
-                break;
-
-            case WSAECONNABORTED:
-                logf("[WS2 HOOK] recv: Connection aborted (WSAECONNABORTED) on socket %u", (unsigned)s);
-                break;
-
-            case WSAENETDOWN:
-                logf("[WS2 HOOK] recv: Network subsystem failure (WSAENETDOWN) on socket %u", (unsigned)s);
-                break;
-
-            case WSAETIMEDOUT:
-                logf("[WS2 HOOK] recv: Connection timeout (WSAETIMEDOUT) on socket %u", (unsigned)s);
-                break;
-
-            case WSAENOTCONN:
-                logf("[WS2 HOOK] recv: Socket not connected (WSAENOTCONN) on socket %u", (unsigned)s);
-                break;
-
-            default:
-                logf("[WS2 HOOK] recv: Error %d (%s) on socket %u",
-                     error, GetErrorDescription(error), (unsigned)s);
-                break;
+        if (error == WSAEWOULDBLOCK) {
+            // Convert WSAEWOULDBLOCK to 0 for server.dll calls
+            logf("[WS2 HOOK] recv: WSAEWOULDBLOCK converted to 0 (server.dll caller)");
+            WSASetLastError(NO_ERROR);
+            return 0;
         }
+
+        log_winsock_error("[WS2 HOOK] recv", s, error);
     } else if (result == 0) {
         logf("[WS2 HOOK] recv: Connection gracefully closed by peer on socket %u", (unsigned)s);
     } else {
@@ -283,32 +184,20 @@ int WINAPI hook_send(SOCKET s, const char *buf, int len, int flags) {
 
         if (sent == SOCKET_ERROR) {
             int error = WSAGetLastError();
-            switch (error) {
-                case WSAEWOULDBLOCK:
-                    logf("[WS2 HOOK] send: WSAEWOULDBLOCK, retrying... (%d/%d)",
-                         retry_count + 1, MAX_RETRIES);
-                    Sleep(1);
-                    retry_count++;
-                    continue;
-
-                case WSAECONNRESET:
-                    logf("[WS2 HOOK] send: Connection reset by peer after sending %d/%d bytes",
-                         total, len);
-                    WSASetLastError(error);
-                    return total > 0 ? total : SOCKET_ERROR;
-
-                case WSAECONNABORTED:
-                    logf("[WS2 HOOK] send: Connection aborted after sending %d/%d bytes",
-                         total, len);
-                    WSASetLastError(error);
-                    return total > 0 ? total : SOCKET_ERROR;
-
-                default:
-                    logf("[WS2 HOOK] send: Fatal error %d (%s) after sending %d/%d bytes",
-                         error, GetErrorDescription(error), total, len);
-                    WSASetLastError(error);
-                    return SOCKET_ERROR;
+            if (error == WSAEWOULDBLOCK) {
+                logf("[WS2 HOOK] send: WSAEWOULDBLOCK, retrying... (%d/%d)",
+                     retry_count + 1, MAX_RETRIES);
+                Sleep(1);
+                retry_count++;
+                continue;
             }
+
+            log_winsock_error("[WS2 HOOK] send", s, error);
+            WSASetLastError(error);
+            if (error == WSAECONNRESET || error == WSAECONNABORTED) {
+                return total > 0 ? total : SOCKET_ERROR;
+            }
+            return SOCKET_ERROR;
         }
 
         if (sent == 0) {
