+++ src/utils.h
@@ -1,0 +1,7 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <winsock2.h>
+const char* GetErrorDescription(int errorCode);
+
+#endif // UTILS_H
